requirements.txt
chardet==5.2.0
joblib==1.4.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
plotly==5.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
tenacity==9.0.0
threadpoolctl==3.5.0
tzdata==2024.1
